chore: rename `WindowsWorker` to `GenericWorker`
It is capable of running on linux with close to feature parity
