+++ shared/packages/worker/src/worker/workers/genericWorker/expectationHandlers/mediaFileThumbnail.ts
@@ -1,5 +1,3 @@
-import { BaseWorker } from '../../../worker'
-import { getStandardCost } from '../lib/lib'
 import {
 	Accessor,
 	hashObj,
@@ -13,6 +11,8 @@ import {
 	stringifyError,
 	startTimer,
 } from '@sofie-package-manager/api'
+import { getStandardCost } from '../lib/lib'
+import { BaseWorker } from '../../../worker'
 import {
 	isFileShareAccessorHandle,
 	isHTTPAccessorHandle,
@@ -22,20 +22,21 @@ import {
 import { IWorkInProgress, WorkInProgress } from '../../../lib/workInProgress'
 import {
 	checkWorkerHasAccessToPackageContainersOnPackage,
+	formatTimeCode,
 	lookupAccessorHandles,
 	LookupPackageContainer,
-	previewFFMpegArguments,
+	thumbnailFFMpegArguments,
 } from './lib'
 import { FFMpegProcess, spawnFFMpeg } from './lib/ffmpeg'
-import { ExpectationWindowsHandler, WindowsWorker } from '../windowsWorker'
-import { scanWithFFProbe, FFProbeScanResult } from './lib/scan'
+import { ExpectationHandlerGenericWorker, GenericWorker } from '../genericWorker'
 import { CancelablePromise } from '../../../lib/cancelablePromise'
+import { scanWithFFProbe, FFProbeScanResult } from './lib/scan'
 
 /**
- * Generates a low-res preview video of a source video file, and stores the resulting file into the target PackageContainer
+ * Generates a thumbnail image from a source video file, and stores the resulting file into the target PackageContainer
  */
-export const MediaFilePreview: ExpectationWindowsHandler = {
-	doYouSupportExpectation(exp: Expectation.Any, worker: WindowsWorker): ReturnTypeDoYouSupportExpectation {
+export const MediaFileThumbnail: ExpectationHandlerGenericWorker = {
+	doYouSupportExpectation(exp: Expectation.Any, worker: GenericWorker): ReturnTypeDoYouSupportExpectation {
 		if (worker.testFFMpeg)
 			return {
 				support: false,
@@ -54,25 +55,24 @@ export const MediaFilePreview: ExpectationWindowsHandler = {
 			}
 		return checkWorkerHasAccessToPackageContainersOnPackage(worker, {
 			sources: exp.startRequirement.sources,
-			targets: exp.endRequirement.targets,
 		})
 	},
 	getCostForExpectation: async (
 		exp: Expectation.Any,
 		worker: BaseWorker
 	): Promise<ReturnTypeGetCostFortExpectation> => {
-		if (!isMediaFilePreview(exp)) throw new Error(`Wrong exp.type: "${exp.type}"`)
+		if (!isMediaFileThumbnail(exp)) throw new Error(`Wrong exp.type: "${exp.type}"`)
 		return getStandardCost(exp, worker)
 	},
 	isExpectationReadyToStartWorkingOn: async (
 		exp: Expectation.Any,
 		worker: BaseWorker
 	): Promise<ReturnTypeIsExpectationReadyToStartWorkingOn> => {
-		if (!isMediaFilePreview(exp)) throw new Error(`Wrong exp.type: "${exp.type}"`)
+		if (!isMediaFileThumbnail(exp)) throw new Error(`Wrong exp.type: "${exp.type}"`)
 
-		const lookupSource = await lookupPreviewSources(worker, exp)
+		const lookupSource = await lookupThumbnailSources(worker, exp)
 		if (!lookupSource.ready) return { ready: lookupSource.ready, sourceExists: false, reason: lookupSource.reason }
-		const lookupTarget = await lookupPreviewTargets(worker, exp)
+		const lookupTarget = await lookupThumbnailTargets(worker, exp)
 		if (!lookupTarget.ready) return { ready: lookupTarget.ready, reason: lookupTarget.reason }
 
 		const tryReading = await lookupSource.handle.tryPackageRead()
@@ -88,9 +88,9 @@ export const MediaFilePreview: ExpectationWindowsHandler = {
 		_wasFulfilled: boolean,
 		worker: BaseWorker
 	): Promise<ReturnTypeIsExpectationFulfilled> => {
-		if (!isMediaFilePreview(exp)) throw new Error(`Wrong exp.type: "${exp.type}"`)
+		if (!isMediaFileThumbnail(exp)) throw new Error(`Wrong exp.type: "${exp.type}"`)
 
-		const lookupSource = await lookupPreviewSources(worker, exp)
+		const lookupSource = await lookupThumbnailSources(worker, exp)
 		if (!lookupSource.ready)
 			return {
 				fulfilled: false,
@@ -99,7 +99,7 @@ export const MediaFilePreview: ExpectationWindowsHandler = {
 					tech: `Not able to access source: ${lookupSource.reason.tech}`,
 				},
 			}
-		const lookupTarget = await lookupPreviewTargets(worker, exp)
+		const lookupTarget = await lookupThumbnailTargets(worker, exp)
 		if (!lookupTarget.ready)
 			return {
 				fulfilled: false,
@@ -127,91 +127,91 @@ export const MediaFilePreview: ExpectationWindowsHandler = {
 		if (!metadata) {
 			return {
 				fulfilled: false,
-				reason: { user: `The preview needs to be re-generated`, tech: `No preview metadata file found` },
+				reason: { user: `The thumbnail needs to be re-generated`, tech: `No thumbnail metadata file found` },
 			}
 		} else if (metadata.sourceVersionHash !== actualSourceVersionHash) {
 			return {
 				fulfilled: false,
 				reason: {
-					user: `The preview needs to be re-generated`,
-					tech: `Preview version doesn't match source file`,
+					user: `The thumbnail needs to be re-generated`,
+					tech: `Thumbnail version doesn't match thumbnail file`,
 				},
 			}
 		} else {
-			return {
-				fulfilled: true,
-				// reason: { user: `Preview already matches preview file`, tech: `Preview already matches preview file` },
-			}
+			return { fulfilled: true }
 		}
 	},
 	workOnExpectation: async (exp: Expectation.Any, worker: BaseWorker): Promise<IWorkInProgress> => {
-		if (!isMediaFilePreview(exp)) throw new Error(`Wrong exp.type: "${exp.type}"`)
-		// Copies the file from Source to Target
+		if (!isMediaFileThumbnail(exp)) throw new Error(`Wrong exp.type: "${exp.type}"`)
+		// Create a thumbnail from the source media file
 
 		const timer = startTimer()
 
-		const lookupSource = await lookupPreviewSources(worker, exp)
+		const lookupSource = await lookupThumbnailSources(worker, exp)
 		if (!lookupSource.ready) throw new Error(`Can't start working due to source: ${lookupSource.reason.tech}`)
 
-		const lookupTarget = await lookupPreviewTargets(worker, exp)
+		const lookupTarget = await lookupThumbnailTargets(worker, exp)
 		if (!lookupTarget.ready) throw new Error(`Can't start working due to target: ${lookupTarget.reason.tech}`)
 
-		const sourceHandle = lookupSource.handle
-		const targetHandle = lookupTarget.handle
-
-		if (
-			(lookupSource.accessor.type === Accessor.AccessType.LOCAL_FOLDER ||
-				lookupSource.accessor.type === Accessor.AccessType.FILE_SHARE ||
-				lookupSource.accessor.type === Accessor.AccessType.HTTP ||
-				lookupSource.accessor.type === Accessor.AccessType.HTTP_PROXY) &&
-			(lookupTarget.accessor.type === Accessor.AccessType.LOCAL_FOLDER ||
-				lookupTarget.accessor.type === Accessor.AccessType.FILE_SHARE ||
-				lookupTarget.accessor.type === Accessor.AccessType.HTTP_PROXY)
-		) {
-			// We can read the source and write the preview directly.
+		let ffMpegProcess: FFMpegProcess | undefined
+		let ffProbeProcess: CancelablePromise<any> | undefined
+		const workInProgress = new WorkInProgress({ workLabel: 'Generating thumbnail' }, async () => {
+			// On cancel
+			ffMpegProcess?.cancel()
+			ffProbeProcess?.cancel()
+		}).do(async () => {
 			if (
-				!isLocalFolderAccessorHandle(sourceHandle) &&
-				!isFileShareAccessorHandle(sourceHandle) &&
-				!isHTTPAccessorHandle(sourceHandle) &&
-				!isHTTPProxyAccessorHandle(sourceHandle)
-			)
-				throw new Error(`Source AccessHandler type is wrong`)
-			if (
-				!isLocalFolderAccessorHandle(targetHandle) &&
-				!isFileShareAccessorHandle(targetHandle) &&
-				!isHTTPProxyAccessorHandle(targetHandle)
-			)
-				throw new Error(`Target AccessHandler type is wrong`)
+				(lookupSource.accessor.type === Accessor.AccessType.LOCAL_FOLDER ||
+					lookupSource.accessor.type === Accessor.AccessType.FILE_SHARE ||
+					lookupTarget.accessor.type === Accessor.AccessType.HTTP ||
+					lookupTarget.accessor.type === Accessor.AccessType.HTTP_PROXY) &&
+				(lookupTarget.accessor.type === Accessor.AccessType.LOCAL_FOLDER ||
+					lookupTarget.accessor.type === Accessor.AccessType.FILE_SHARE ||
+					lookupTarget.accessor.type === Accessor.AccessType.HTTP_PROXY)
+			) {
+				const sourceHandle = lookupSource.handle
+				const targetHandle = lookupTarget.handle
+				if (
+					!isLocalFolderAccessorHandle(sourceHandle) &&
+					!isHTTPAccessorHandle(sourceHandle) &&
+					!isFileShareAccessorHandle(sourceHandle) &&
+					!isHTTPProxyAccessorHandle(sourceHandle)
+				)
+					throw new Error(`Source AccessHandler type is wrong`)
+				if (
+					!isLocalFolderAccessorHandle(targetHandle) &&
+					!isFileShareAccessorHandle(targetHandle) &&
+					!isHTTPProxyAccessorHandle(targetHandle)
+				)
+					throw new Error(`Target AccessHandler type is wrong`)
 
-			let ffMpegProcess: FFMpegProcess | undefined
-			let ffProbeProcess: CancelablePromise<any> | undefined
-			const workInProgress = new WorkInProgress({ workLabel: 'Generating preview' }, async () => {
-				// On cancel
-				ffMpegProcess?.cancel()
-				ffProbeProcess?.cancel()
-			}).do(async () => {
 				const tryReadPackage = await sourceHandle.checkPackageReadAccess()
-				if (!tryReadPackage.success) throw new Error(tryReadPackage.reason.tech)
+				if (!tryReadPackage.success) {
+					throw new Error(tryReadPackage.reason.tech)
+				}
 
 				const actualSourceVersion = await sourceHandle.getPackageActualVersion()
-				const actualSourceVersionHash = hashObj(actualSourceVersion)
-				// const actualSourceUVersion = makeUniversalVersion(actualSourceVersion)
+				const sourceVersionHash = hashObj(actualSourceVersion)
 
 				const metadata: Metadata = {
-					sourceVersionHash: actualSourceVersionHash,
+					sourceVersionHash: sourceVersionHash,
 					version: {
 						...{
 							// Default values:
-							type: Expectation.Version.Type.MEDIA_FILE_PREVIEW,
-							bitrate: '40k',
-							width: 320,
+							type: Expectation.Version.Type.MEDIA_FILE_THUMBNAIL,
+							width: 256,
 							height: -1,
+							seekTime: 0,
 						},
 						...exp.endRequirement.version,
 					},
 				}
 
-				await targetHandle.removePackage('Prepare for preview generation')
+				await targetHandle.removePackage('Prepare for thumbnail generation')
+
+				const seekTime = exp.endRequirement.version.seekTime
+
+				const seekTimeCode: string | undefined = seekTime !== undefined ? formatTimeCode(seekTime) : undefined
 
 				let inputPath: string
 				if (isLocalFolderAccessorHandle(sourceHandle)) {
@@ -234,38 +234,28 @@ export const MediaFilePreview: ExpectationWindowsHandler = {
 				ffProbeProcess = undefined
 				const hasVideoStream =
 					ffProbeScan.streams && ffProbeScan.streams.some((stream) => stream.codec_type === 'video')
-				if (!hasVideoStream) {
-					workInProgress._reportComplete(
-						actualSourceVersionHash,
-						{
-							user: `Preview generation skipped due to file having no video streams`,
-							tech: `Completed at ${Date.now()}`,
-						},
-						undefined
-					)
-					return
-				}
 
-				const args = previewFFMpegArguments(inputPath, true, metadata)
+				// Use FFMpeg to generate the thumbnail:
+				const args = thumbnailFFMpegArguments(inputPath, metadata, seekTimeCode, hasVideoStream)
 
-				const fileOperation = await targetHandle.prepareForOperation('Generate preview', lookupSource.handle)
+				const fileOperation = await targetHandle.prepareForOperation('Generate thumbnail', lookupSource.handle)
 
 				ffMpegProcess = await spawnFFMpeg(
 					args,
 					targetHandle,
-					// actualSourceVersionHash,
 					async () => {
 						// Called when ffmpeg has finished
 						worker.logger.debug(`FFMpeg finished [PID=${ffMpegProcess?.pid}]: ${args.join(' ')}`)
 						ffMpegProcess = undefined
+
 						await targetHandle.finalizePackage(fileOperation)
 						await targetHandle.updateMetadata(metadata)
 
 						const duration = timer.get()
 						workInProgress._reportComplete(
-							actualSourceVersionHash,
+							sourceVersionHash,
 							{
-								user: `Preview generation completed in ${Math.round(duration / 100) / 10}s`,
+								user: `Thumbnail generation completed in ${Math.round(duration / 100) / 10}s`,
 								tech: `Completed at ${Date.now()}`,
 							},
 							undefined
@@ -279,25 +269,23 @@ export const MediaFilePreview: ExpectationWindowsHandler = {
 						workInProgress._reportError(err)
 					},
 					async (progress: number) => {
-						workInProgress._reportProgress(actualSourceVersionHash, progress)
+						workInProgress._reportProgress(sourceVersionHash, progress)
 					}
-					//,worker.logger.debug
+					// ,worker.logger.debug
 				)
 				worker.logger.debug(`FFMpeg started [PID=${ffMpegProcess.pid}]: ${args.join(' ')}`)
-			})
+			} else {
+				throw new Error(
+					`MediaFileThumbnail.workOnExpectation: Unsupported accessor source-target pair "${lookupSource.accessor.type}"-"${lookupTarget.accessor.type}"`
+				)
+			}
+		})
 
-			return workInProgress
-		} else {
-			throw new Error(
-				`MediaFilePreview.workOnExpectation: Unsupported accessor source-target pair "${lookupSource.accessor.type}"-"${lookupTarget.accessor.type}"`
-			)
-		}
+		return workInProgress
 	},
 	removeExpectation: async (exp: Expectation.Any, worker: BaseWorker): Promise<ReturnTypeRemoveExpectation> => {
-		if (!isMediaFilePreview(exp)) throw new Error(`Wrong exp.type: "${exp.type}"`)
-		// Remove the file on the location
-
-		const lookupTarget = await lookupPreviewTargets(worker, exp)
+		if (!isMediaFileThumbnail(exp)) throw new Error(`Wrong exp.type: "${exp.type}"`)
+		const lookupTarget = await lookupThumbnailTargets(worker, exp)
 		if (!lookupTarget.ready) {
 			return {
 				removed: false,
@@ -320,24 +308,21 @@ export const MediaFilePreview: ExpectationWindowsHandler = {
 			}
 		}
 
-		return {
-			removed: true,
-			// reason: { user: ``, tech: `Removed preview file "${exp.endRequirement.content.filePath}" from target` },
-		}
+		return { removed: true }
 	},
 }
-function isMediaFilePreview(exp: Expectation.Any): exp is Expectation.MediaFilePreview {
-	return exp.type === Expectation.Type.MEDIA_FILE_PREVIEW
+function isMediaFileThumbnail(exp: Expectation.Any): exp is Expectation.MediaFileThumbnail {
+	return exp.type === Expectation.Type.MEDIA_FILE_THUMBNAIL
 }
 
 interface Metadata {
 	sourceVersionHash: string
-	version: Expectation.Version.MediaFilePreview
+	version: Expectation.Version.MediaFileThumbnail
 }
 
-async function lookupPreviewSources(
+async function lookupThumbnailSources(
 	worker: BaseWorker,
-	exp: Expectation.MediaFilePreview
+	exp: Expectation.MediaFileThumbnail
 ): Promise<LookupPackageContainer<Metadata>> {
 	return lookupAccessorHandles<Metadata>(
 		worker,
@@ -351,9 +336,9 @@ async function lookupPreviewSources(
 		}
 	)
 }
-async function lookupPreviewTargets(
+async function lookupThumbnailTargets(
 	worker: BaseWorker,
-	exp: Expectation.MediaFilePreview
+	exp: Expectation.MediaFileThumbnail
 ): Promise<LookupPackageContainer<Metadata>> {
 	return lookupAccessorHandles<Metadata>(
 		worker,
