+++ shared/packages/worker/src/worker/workers/genericWorker/expectationHandlers/packageScan.ts
@@ -15,24 +15,16 @@ import {
 import { isCorePackageInfoAccessorHandle } from '../../../accessorHandlers/accessor'
 import { IWorkInProgress, WorkInProgress } from '../../../lib/workInProgress'
 import { checkWorkerHasAccessToPackageContainersOnPackage, lookupAccessorHandles, LookupPackageContainer } from './lib'
-import { DeepScanResult, FieldOrder, PackageInfoType, ScanAnomaly } from './lib/coreApi'
 import { CancelablePromise } from '../../../lib/cancelablePromise'
-import {
-	FFProbeScanResult,
-	isAnFFMpegSupportedSourceAccessor,
-	isAnFFMpegSupportedSourceAccessorHandle,
-	scanFieldOrder,
-	scanMoreInfo,
-	scanWithFFProbe,
-} from './lib/scan'
-import { ExpectationWindowsHandler, WindowsWorker } from '../windowsWorker'
+import { isAnFFMpegSupportedSourceAccessor, isAnFFMpegSupportedSourceAccessorHandle, scanWithFFProbe } from './lib/scan'
+import { ExpectationHandlerGenericWorker, GenericWorker } from '../genericWorker'
+import { PackageInfoType } from './lib/coreApi'
 
 /**
- * Performs a "deep scan" of the source package and saves the result file into the target PackageContainer (a Sofie Core collection)
- * The "deep scan" differs from the usual scan in that it does things that takes a bit longer, like scene-detection, field order detection etc..
+ * Scans the source package and saves the result file into the target PackageContainer (a Sofie Core collection)
  */
-export const PackageDeepScan: ExpectationWindowsHandler = {
-	doYouSupportExpectation(exp: Expectation.Any, worker: WindowsWorker): ReturnTypeDoYouSupportExpectation {
+export const PackageScan: ExpectationHandlerGenericWorker = {
+	doYouSupportExpectation(exp: Expectation.Any, worker: GenericWorker): ReturnTypeDoYouSupportExpectation {
 		if (worker.testFFMpeg)
 			return {
 				support: false,
@@ -55,21 +47,21 @@ export const PackageDeepScan: ExpectationWindowsHandler = {
 	},
 	getCostForExpectation: async (
 		exp: Expectation.Any,
-		worker: BaseWorker
+		worker: GenericWorker
 	): Promise<ReturnTypeGetCostFortExpectation> => {
-		if (!isPackageDeepScan(exp)) throw new Error(`Wrong exp.type: "${exp.type}"`)
+		if (!isPackageScan(exp)) throw new Error(`Wrong exp.type: "${exp.type}"`)
 		return getStandardCost(exp, worker)
 	},
 
 	isExpectationReadyToStartWorkingOn: async (
 		exp: Expectation.Any,
-		worker: BaseWorker
+		worker: GenericWorker
 	): Promise<ReturnTypeIsExpectationReadyToStartWorkingOn> => {
-		if (!isPackageDeepScan(exp)) throw new Error(`Wrong exp.type: "${exp.type}"`)
+		if (!isPackageScan(exp)) throw new Error(`Wrong exp.type: "${exp.type}"`)
 
-		const lookupSource = await lookupDeepScanSources(worker, exp)
+		const lookupSource = await lookupScanSources(worker, exp)
 		if (!lookupSource.ready) return { ready: lookupSource.ready, sourceExists: false, reason: lookupSource.reason }
-		const lookupTarget = await lookupDeepScanSources(worker, exp)
+		const lookupTarget = await lookupScanSources(worker, exp)
 		if (!lookupTarget.ready) return { ready: lookupTarget.ready, reason: lookupTarget.reason }
 
 		const tryReading = await lookupSource.handle.tryPackageRead()
@@ -85,9 +77,9 @@ export const PackageDeepScan: ExpectationWindowsHandler = {
 		wasFulfilled: boolean,
 		worker: BaseWorker
 	): Promise<ReturnTypeIsExpectationFulfilled> => {
-		if (!isPackageDeepScan(exp)) throw new Error(`Wrong exp.type: "${exp.type}"`)
+		if (!isPackageScan(exp)) throw new Error(`Wrong exp.type: "${exp.type}"`)
 
-		const lookupSource = await lookupDeepScanSources(worker, exp)
+		const lookupSource = await lookupScanSources(worker, exp)
 		if (!lookupSource.ready)
 			return {
 				fulfilled: false,
@@ -96,7 +88,7 @@ export const PackageDeepScan: ExpectationWindowsHandler = {
 					tech: `Not able to access source: ${lookupSource.reason.tech}`,
 				},
 			}
-		const lookupTarget = await lookupDeepScanTargets(worker, exp)
+		const lookupTarget = await lookupScanTargets(worker, exp)
 		if (!lookupTarget.ready)
 			return {
 				fulfilled: false,
@@ -111,7 +103,7 @@ export const PackageDeepScan: ExpectationWindowsHandler = {
 		if (!isCorePackageInfoAccessorHandle(lookupTarget.handle)) throw new Error(`Target AccessHandler type is wrong`)
 
 		const packageInfoSynced = await lookupTarget.handle.findUnUpdatedPackageInfo(
-			PackageInfoType.DeepScan,
+			PackageInfoType.Scan,
 			exp,
 			exp.startRequirement.content,
 			actualSourceVersion,
@@ -120,31 +112,28 @@ export const PackageDeepScan: ExpectationWindowsHandler = {
 		if (packageInfoSynced.needsUpdate) {
 			if (wasFulfilled) {
 				// Remove the outdated scan result:
-				await lookupTarget.handle.removePackageInfo(PackageInfoType.DeepScan, exp)
+				await lookupTarget.handle.removePackageInfo(PackageInfoType.Scan, exp)
 			}
 			return { fulfilled: false, reason: packageInfoSynced.reason }
 		} else {
 			return { fulfilled: true }
 		}
 	},
-	workOnExpectation: async (
-		exp: Expectation.Any,
-		worker: WindowsWorker,
-		progressTimeout: number
-	): Promise<IWorkInProgress> => {
-		if (!isPackageDeepScan(exp)) throw new Error(`Wrong exp.type: "${exp.type}"`)
-		// Scan the source media file and upload the results to Core
+	workOnExpectation: async (exp: Expectation.Any, worker: BaseWorker): Promise<IWorkInProgress> => {
+		if (!isPackageScan(exp)) throw new Error(`Wrong exp.type: "${exp.type}"`)
+		// Scan the source package and upload the results to Core
 		const timer = startTimer()
 
-		const lookupSource = await lookupDeepScanSources(worker, exp)
+		const lookupSource = await lookupScanSources(worker, exp)
 		if (!lookupSource.ready) throw new Error(`Can't start working due to source: ${lookupSource.reason.tech}`)
 
-		const lookupTarget = await lookupDeepScanTargets(worker, exp)
+		const lookupTarget = await lookupScanTargets(worker, exp)
 		if (!lookupTarget.ready) throw new Error(`Can't start working due to target: ${lookupTarget.reason.tech}`)
 
 		let currentProcess: CancelablePromise<any> | undefined
-		const workInProgress = new WorkInProgress({ workLabel: 'Deep Scanning file' }, async () => {
+		const workInProgress = new WorkInProgress({ workLabel: 'Scanning file' }, async () => {
 			// On cancel
+
 			currentProcess?.cancel()
 		}).do(async () => {
 			const sourceHandle = lookupSource.handle
@@ -154,7 +143,7 @@ export const PackageDeepScan: ExpectationWindowsHandler = {
 				lookupTarget.accessor.type !== Accessor.AccessType.CORE_PACKAGE_INFO
 			)
 				throw new Error(
-					`PackageDeepScan.workOnExpectation: Unsupported accessor source-target pair "${lookupSource.accessor.type}"-"${lookupTarget.accessor.type}"`
+					`PackageScan.workOnExpectation: Unsupported accessor source-target pair "${lookupSource.accessor.type}"-"${lookupTarget.accessor.type}"`
 				)
 
 			if (!isAnFFMpegSupportedSourceAccessorHandle(sourceHandle))
@@ -168,128 +157,23 @@ export const PackageDeepScan: ExpectationWindowsHandler = {
 			const actualSourceVersion = await sourceHandle.getPackageActualVersion()
 			const sourceVersionHash = hashObj(actualSourceVersion)
 
-			workInProgress._reportProgress(sourceVersionHash, 0.01)
+			workInProgress._reportProgress(sourceVersionHash, 0.1)
 
 			// Scan with FFProbe:
 			currentProcess = scanWithFFProbe(sourceHandle)
-			const ffProbeScan: FFProbeScanResult = await currentProcess
-			const hasVideoStream =
-				ffProbeScan.streams && ffProbeScan.streams.some((stream) => stream.codec_type === 'video')
-			workInProgress._reportProgress(sourceVersionHash, 0.1)
+			const scanResult = await currentProcess
+			workInProgress._reportProgress(sourceVersionHash, 0.5)
 			currentProcess = undefined
 
-			// Scan field order:
-			let resultFieldOrder = FieldOrder.Unknown
-			if (hasVideoStream) {
-				currentProcess = scanFieldOrder(sourceHandle, exp.endRequirement.version)
-				resultFieldOrder = await currentProcess
-				currentProcess = undefined
-			}
-			workInProgress._reportProgress(sourceVersionHash, 0.2)
-
-			// Scan more info:
-			let resultBlacks: ScanAnomaly[] = []
-			let resultFreezes: ScanAnomaly[] = []
-			let resultScenes: number[] = []
-			if (hasVideoStream) {
-				let hasGottenProgress = false
-
-				currentProcess = new CancelablePromise<{
-					scenes: number[]
-					freezes: ScanAnomaly[]
-					blacks: ScanAnomaly[]
-				}>(async (resolve, reject, onCancel) => {
-					let isDone = false
-					let ignoreNextCancelError = false
-
-					const scanMoreInfoProcess = scanMoreInfo(
-						sourceHandle,
-						ffProbeScan,
-						exp.endRequirement.version,
-						(progress) => {
-							hasGottenProgress = true
-							workInProgress._reportProgress(sourceVersionHash, 0.21 + 0.77 * progress)
-						},
-						worker.logger.category('scanMoreInfo')
-					)
-					onCancel(() => {
-						scanMoreInfoProcess.cancel()
-					})
-
-					scanMoreInfoProcess.then(
-						(result) => {
-							isDone = true
-							resolve(result)
-						},
-						(error) => {
-							if (`${error}`.match(/cancelled/i) && ignoreNextCancelError) {
-								// ignore this
-								ignoreNextCancelError = false
-							} else {
-								reject(error)
-							}
-						}
-					)
-
-					// Guard against an edge case where we don't get any progress reports:
-					setTimeout(() => {
-						if (!isDone && currentProcess && !hasGottenProgress) {
-							// If we haven't gotten any progress yet, we probably won't get any.
-
-							// 2023-09-20: There seems to be some bug in the FFMpeg scan where it won't output any progress
-							// if the scene detection is on.
-							// Let's abort and try again without scene detection:
-
-							ignoreNextCancelError = true
-							currentProcess.cancel()
-
-							const scanMoreInfoProcessSecondTry = scanMoreInfo(
-								sourceHandle,
-								ffProbeScan,
-								{
-									...exp.endRequirement.version,
-									scenes: false, // no scene detection
-								},
-								(progress) => {
-									hasGottenProgress = true
-									workInProgress._reportProgress(sourceVersionHash, 0.21 + 0.77 * progress)
-								},
-								worker.logger.category('scanMoreInfo')
-							)
-
-							scanMoreInfoProcessSecondTry.then(
-								(result) => resolve(result),
-								(error) => reject(error)
-							)
-						}
-					}, progressTimeout * 0.5)
-				})
-
-				const result = await currentProcess
-
-				resultBlacks = result.blacks
-				resultFreezes = result.freezes
-				resultScenes = result.scenes
-				currentProcess = undefined
-			}
-			workInProgress._reportProgress(sourceVersionHash, 0.99)
-
-			const deepScan: DeepScanResult = {
-				field_order: resultFieldOrder,
-				blacks: resultBlacks,
-				freezes: resultFreezes,
-				scenes: resultScenes,
-			}
-
 			// all done:
-			const scanOperation = await targetHandle.prepareForOperation('Deep scan', sourceHandle)
+			const scanOperation = await targetHandle.prepareForOperation('Scan', sourceHandle)
 			await targetHandle.updatePackageInfo(
-				PackageInfoType.DeepScan,
+				PackageInfoType.Scan,
 				exp,
 				exp.startRequirement.content,
 				actualSourceVersion,
 				exp.endRequirement.version,
-				deepScan
+				scanResult
 			)
 
 			await targetHandle.finalizePackage(scanOperation)
@@ -308,8 +192,8 @@ export const PackageDeepScan: ExpectationWindowsHandler = {
 		return workInProgress
 	},
 	removeExpectation: async (exp: Expectation.Any, worker: BaseWorker): Promise<ReturnTypeRemoveExpectation> => {
-		if (!isPackageDeepScan(exp)) throw new Error(`Wrong exp.type: "${exp.type}"`)
-		const lookupTarget = await lookupDeepScanTargets(worker, exp)
+		if (!isPackageScan(exp)) throw new Error(`Wrong exp.type: "${exp.type}"`)
+		const lookupTarget = await lookupScanTargets(worker, exp)
 		if (!lookupTarget.ready)
 			return {
 				removed: false,
@@ -321,7 +205,7 @@ export const PackageDeepScan: ExpectationWindowsHandler = {
 		if (!isCorePackageInfoAccessorHandle(lookupTarget.handle)) throw new Error(`Target AccessHandler type is wrong`)
 
 		try {
-			await lookupTarget.handle.removePackageInfo(PackageInfoType.DeepScan, exp)
+			await lookupTarget.handle.removePackageInfo(PackageInfoType.Scan, exp)
 		} catch (err) {
 			return {
 				removed: false,
@@ -335,14 +219,14 @@ export const PackageDeepScan: ExpectationWindowsHandler = {
 		return { removed: true }
 	},
 }
-function isPackageDeepScan(exp: Expectation.Any): exp is Expectation.PackageDeepScan {
-	return exp.type === Expectation.Type.PACKAGE_DEEP_SCAN
+function isPackageScan(exp: Expectation.Any): exp is Expectation.PackageScan {
+	return exp.type === Expectation.Type.PACKAGE_SCAN
 }
 type Metadata = any // not used
 
-async function lookupDeepScanSources(
+async function lookupScanSources(
 	worker: BaseWorker,
-	exp: Expectation.PackageDeepScan
+	exp: Expectation.PackageScan
 ): Promise<LookupPackageContainer<Metadata>> {
 	return lookupAccessorHandles<Metadata>(
 		worker,
@@ -356,9 +240,9 @@ async function lookupDeepScanSources(
 		}
 	)
 }
-async function lookupDeepScanTargets(
+async function lookupScanTargets(
 	worker: BaseWorker,
-	exp: Expectation.PackageDeepScan
+	exp: Expectation.PackageScan
 ): Promise<LookupPackageContainer<Metadata>> {
 	return lookupAccessorHandles<Metadata>(
 		worker,
